requirements.txt
plotly
streamlit
supabase
python-dotenv
pandas
